"ads2_m12_a1_1: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS - 2 Assignments/M 12/ADS-2_Exam-2/Edge.java
@@ -37,7 +37,7 @@ public class Edge implements Comparable<Edge> {
     /**
      * weight.
      */
-    private final double weight;
+    private final float weight;
 
     /**
      * Initializes an edge between vertices {@code v} and {@code w} of
@@ -50,7 +50,7 @@ public class Edge implements Comparable<Edge> {
      *         is a negative integer
      * @throws IllegalArgumentException if {@code weight} is {@code NaN}
      */
-    public Edge(final int v1, final int w1, final double weight1) {
+    public Edge(final int v1, final int w1, final float weight1) {
         if (v1 < 0) {
             throw new IllegalArgumentException(
                 "vertex index must be a nonnegative integer");
